gpt_backend/auth_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic
pytest
httpx
passlib[bcrypt]
pyjwt
pyyaml
requests
bcrypt==4.0.1
structlog
python-json-logger
